Integrating Igor's comments.
