+++ tso-server/src/main/java/com/yahoo/omid/tso/PersistenceProcessor.java
@@ -24,6 +24,6 @@ interface PersistenceProcessor {
 
     void persistTimestamp(long startTimestamp, Channel c, MonitoringContext monCtx);
     void persistLowWatermark(long lowWatermark, MonitoringContext monCtx);
-    void persistFlush(boolean forTesting);
+    void persistFlush();
     void reset();
 }
