+++ tso-server/src/test/java/com/yahoo/omid/tso/TSOMockModule.java
@@ -11,6 +11,7 @@ import com.yahoo.omid.tso.TimestampOracleImpl.InMemoryTimestampStorage;
 
 import javax.inject.Named;
 import javax.inject.Singleton;
+
 import java.net.SocketException;
 import java.net.UnknownHostException;
